Update test core to allow for true/false and skips
This allows the addition of a test suite that runs a command and expects
the result to be true (return zero) or false (return non-zero). This
also adds the facility to skip a test if certain criteria are not met.

This commit also merges the test runner core to main run_test.sh script,
thereby allowing for easier edits to the test core.

Finally, this ensures that the test runner changes the working directory
to the bashlib root directory. While bashlib itself can find the
relative paths to the individual modules, we need to have the right link
to the bashlib top-level script.